adds item-list, price-list (+icon), these use a DoubleEditView for entering data, this view adds to form, persists data and saves to db. Common view functions refactored to ViewUtils

+++ app/src/main/java/com/sundown/maplists/activities/LocationListActivity.java
@@ -184,8 +184,10 @@ public class LocationListActivity extends AppCompatActivity implements ActionDia
 
     @Override
     public void confirmAction(boolean confirmed) {
-        db.delete(model.getDocumentId(), model.getMapId(), Operation.DELETE_LOCATION_LIST);
-        returnActivityResult();
+        if (confirmed) {
+            db.delete(model.getDocumentId(), model.getMapId(), Operation.DELETE_LOCATION_LIST);
+            returnActivityResult();
+        }
     }
 
     private void returnActivityResult(){
